+++ CHANGELOG.md
@@ -1,4 +1,4 @@
-# Next version
+# 0.3.0
 + Check if public IP is different from DNS record
 + Add browse zones test
 + Fix code smells